Migrate components/panels
